+++ src/Employer/Employer.Web/Configuration/ConfigurationExtensions.cs
@@ -78,6 +78,7 @@ namespace Esfa.Recruit.Employer.Web.Configuration
                 }
 
                 opts.Filters.Add(new AutoValidateAntiforgeryTokenAttribute());
+                opts.Filters.Add(typeof(LevyDeclarationCheckFilter), 50);
 
                 if (EnvironmentNames.GetTestEnvironmentNames().Contains(hostingEnvironment.EnvironmentName) == false)
                 {
