+++ src/Employer/Employer.Web/Configuration/DataProtectionPurposes.cs
@@ -1,0 +1,8 @@
+
+namespace Esfa.Recruit.Employer.Web.Configuration
+{
+    public static class DataProtectionPurposes
+    {
+        public const string LevyDeclarationCookie = "levy-declaration";
+    }
+}
